mmt-app-frontline-services/main/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.1.0